Fix buttons intersecting on Subscribed Channels tab in many languages
Note that this can still occur if one of the words in 'Unsubscribe' is pretty big on smaller screen sizes, e.g., Svenska at 515px. This is because we allow the Unsubscribe button to be wider than the element itself, which is hard to change here without completely reworking the structure using  (due to limitations with minmax being unable to have fit-content in a grid definition).