Implement Watch Later
Watch Later is now protected, given that its manipulation requires its specific id. The addToPlaylists icon is now persistently visible whenever the favorites icon is (i.e., when the video is favorited). The :focus-within logic to that was changed to be :has(:focus-visible) to fix an issue with it staying visible after clicked (pre-existing).