Settings menu & UX improvements (#5029)
* Implement settings section sorting

* Make IS_ELECTRON property used directly

* Implement settings menu with icons

* Implement short labels and fallback to longer label when translation doesn't exist

* Implement dynamic scroll-based underlining of active section

* Ensure General Settings is always first, ignoring sorting

Solving the problem of preventing 'jumping around' when changing locales with alphabetical sorting is hard. This problem is easily solved by stickying general settings to the top of the list unconditionally. This is in line with ChunkyProgrammer's initial assessment of how to improve the settings order in https://github.com/FreeTubeApp/FreeTube/issues/1739, albeit not also moving Theme Settings to the top for the time being. The rationale from a functional level is that General Settings is a hub. Even when you change languages, change sort order, or what have you, General Settings is right at the top. I don't imagine we need to update the label of the setting, as I think this relationship is quite intuitive.

* Styling & documentation improvements; remove problematized Hot Pink link underline removal experiment

* Implement bare minimum mobile styling

* Increase link padding

* Fix underline to be equal on both sides

* Update to have first section active by default

* Make constant for 'active' class name

* Replace proxy icon with more fitting one

* Move shortTitle logic

* Update scrolling logic to not use anchor links

* Implement mobile view for Settings Menu

* Replace rounded corners and move section titles to the interior

* Update font sizes and breakpoints to work for all supported devices & languages

* Update as per review comments

* Make class-specific constants declared in file

* Update settings menu link sizing to be dynamic to better accommodate mobile/tablet devices

* Focus active menu on mobile close

* Replace shortTitle with title, & replace EN-US labels with shorter forms

* Update menu font size for mobile devices

* Focus section heading on navigating to it in desktop view

---------

Co-authored-by: Jason <84899178+jasonhenriquez@users.noreply.github.com>