Partial fix for playlist title and description overflow
This change means that text-ellipsis no longer works, but it looks like that might have been a mistake. Because when it worked correctly, there was no way for the user to read the full playlist title when it was too long. Now the title and description text will wrap properly and not be cut off. Unfortunately this pushes the search bar to the bottom and causes it to be partially cut off, so that has to be fixed