Add Everforest color themes (#6468)
* feat: add everforest color scheme

Adds the Everforest color scheme

* perf: compress svg file sizes

This compresses the svg file sizes by using https://svgoptimizer.com

* fix: fix light secondary accent colors

Fixes the light secondary accent colors.
Previously they were only three letters, which led to unmatching, ugly looking colors.

* refactor: rename turquoise to aqua

Renames the turquoise accent color to aqua, matching the official name of the palette.

* refactor: keep a-z sorting in themesettings

This keeps the a-z sorting in the ThemeSettings.vue file

* fix: change missing css from turquoise to aqua

Changes the css I have missed from turquoise to aqua

* feat: apply a-z sorting in color.js files

This also applies the a-z sorting in the colors.js files.

* fix: change light low card bg color for contrast ratio

Modifies the Everforest Light Low bg-color to a lighter value, so the contrast ratio matches the required 4.5+

* Rename missed Turquoise to Aqua

Co-authored-by: efb4f5ff-1298-471a-8973-3d47447115dc <73130443+efb4f5ff-1298-471a-8973-3d47447115dc@users.noreply.github.com>

* fix: add missing css

* fix: add icons to missing css

This adds the logo icons to the "missing css" which is found
when the setting "match top bar with main color" is enabled

* fix: correct light color ordering

Corrects the light color ordering from dark -> light (instead of light -> dark)

* fix: improve contrast ratio

* feat: light accent color contrast

Fixed the colorscheme with the light theme accent colors, by decreasing their lightness.

* fix: update colors in colors.js

Updates the colors in the colors.js file to match the ones in the colors.css file

* fix: make index.js use the correct bg colors

* fix: make medium dark nav color match theme

Makes the medium dark nav bar color match the other theme variants

* Fix themes.css side-nav-color

Co-authored-by: efb4f5ff-1298-471a-8973-3d47447115dc <73130443+efb4f5ff-1298-471a-8973-3d47447115dc@users.noreply.github.com>

---------

Co-authored-by: DontBlameMe <32d1938e-18da-413a-853a-ccec6ded1347@slmails.com>
Co-authored-by: efb4f5ff-1298-471a-8973-3d47447115dc <73130443+efb4f5ff-1298-471a-8973-3d47447115dc@users.noreply.github.com>
Co-authored-by: DontBlameMe <dontblameme@noreply.codeberg.org>