feat: add option to hide the channel link (#7514)
* feat: add option to hide the channel link

Adds an option to hide the channel link from a video

closes #3193

* fix: Do not move sharing actions when channel actions are hidden

This prevents the sharing actions from being moved to where the profile would be,
once the profile is hidden via the parental controls.

* refactor: only hide channel name/logo

This change makes it so only the channel name/icon get hidden, not the subscribe button itself.
This makes sense because there is a separate setting to achive this behavior, if needed.

* refactor: move toggle to the right

Moves the toggle to the right side (grouping in pairs of two),
instead of three on one side and one on the other side

* refactor: rename to hide uploader

* Update static/locales/en-US.yaml

Co-authored-by: efb4f5ff-1298-471a-8973-3d47447115dc <73130443+efb4f5ff-1298-471a-8973-3d47447115dc@users.noreply.github.com>

* Update src/renderer/components/ParentalControlSettings.vue

Co-authored-by: efb4f5ff-1298-471a-8973-3d47447115dc <73130443+efb4f5ff-1298-471a-8973-3d47447115dc@users.noreply.github.com>

* fix: subscribe button under channel text

* fix: linting warnings/indentation

* fix: new linting warnings

---------

Co-authored-by: DontBlameMe <dontblameme@noreply.codeberg.org>
Co-authored-by: efb4f5ff-1298-471a-8973-3d47447115dc <73130443+efb4f5ff-1298-471a-8973-3d47447115dc@users.noreply.github.com>